add file
Create a dashboard for analyzing cash flow on the project. With its help, (imaginary) product managers will track the dynamics of changes in cash flow and make a top-level analysis of the factors behind these changes.
The dashboard should include all or at least most of the following metrics:
Monthly Recurring Revenue (MRR) is the amount of revenue for a calendar month, but only if this revenue was received from sources that recur from month to month.
Paid Users - the number of paying users.
Average Revenue Per Paid User (ARPPU) - the average revenue per paying user. It is calculated as ARPPU = Revenue / Paid Users.
New Paid Users - the number of users who started paying in the corresponding period of time.
New MRR is the MRR generated by users who became paying users in the corresponding month.
Churned Users is the number of users who stopped paying in the corresponding period.
Churn Rate is the ratio of Churned Users in the target period to Paid Users in the previous period. For example, Churn Rate(June) = Churned Users(June) / Paid users(May)
Churned revenue is the total revenue for the previous period from all users who stopped paying.
Revenue Churn rate is the ratio of revenue for the previous month from those users who did not continue to pay to the revenue for a certain calendar month. For example, Revenue Churn Rate (June) = Churned Revenue (June) / MRR (May)
MRR Expansion is the amount by which MRR increased from one month to the next. It is calculated for users who started paying more in the current month.
Contraction MRR is the amount by which MRR decreased from one month to the next. It is calculated for users who started paying less in the current month.
Customer LifeTime (LT) is the average amount of time from the first payment to the user's dropout. That is, the entire time that an average user stays with the product.
Customer LifeTime Value (LTV) is the average total amount that one user pays for the entire time of using the product. That is, all the money that an average user pays to the company.
A dashboard should:
consist of at least five charts,
include filters by date, user language, and user age.
Among the charts, there must be at least two that show the factors that change revenue and paid users from month to month.
Use Tableau for visualization. 
The completed task should consist of two parts:
SQL query with calculation of all metrics. A dashboard in Tableau Public Paste the link to the report into the text field for handing in your homework.